+++ src/utils/jwt.helper.ts
@@ -1,4 +1,4 @@
-// import httpErrors from 'http-errors'
+import httpErrors from 'http-errors'
 import jwt from 'jsonwebtoken'
 
 const signAccessToken = (id: string): Promise<unknown> => {
@@ -6,15 +6,18 @@ const signAccessToken = (id: string): Promise<unknown> => {
     const payload = {
       name: 'Yours truly'
     }
-    const secret = 'Some super secret'
+    const secret = process.env.ACCESS_TOKEN_SECRET as string
+
     const options = {
       audience : id,
       expiresIn: '10m',
       issuer   : process.env.RHOADES_FRONT_URL as string
     }
     jwt.sign(payload, secret, options, (error, token): void => {
-      if (error) reject(error)
-      else resolve(token)
+      if (error) {
+        console.error(error)
+        reject(new httpErrors.InternalServerError('Ups! Something went wrong'))
+      } else resolve(token)
     })
   })
 }